Fixes a bug with the bind operator
The bind operator wasn't able to remember if it has delivered its
mappings between two quantums (when the source iterator is an EmptyIterator).
Consequently it was possible to get into an infinite loop...

To fix this bug, the flag "delivered" is saved into the plan at the end of quantums.