Property paths parser
For the moment, property paths are evaluated out of the scope of
any BGPs. A possible optimization would be to consider the context
of path patterns, especially when property paths can be rewritten
into classical BGPs...

Because there is no cardinality estimator for the property paths,
we use an heuristic to compute the join ordering on BGPs that
contain path patterns.
- Path patterns for which the subject or the object is a constant
  are evaluated first (to avoid multiple calls to the transitive
  closure algorithm)
- Path patterns for which neither the subject nor the object is a
  constant are evaluated last (to bind the variables first)