Detection of duplicates generated by the property paths
Transitive closure expressions may generate duplicates...
To allow the client to detect and remove these duplicates, the server
compute the imprint of each solution mapping and bind it to the variable '?imprint'.
Two solution mappings that have the same value for the variable '?imprint'
are duplicates and one of them must be removed.